fix: remove buggy try except when appending to relationships
It seems that when an exception is raised, the serialization logic of the
exception message leads to an incorrect initialization of the SQLAlchemy
collection to a raw empty list instead of an instrumented list or a
QueryAppender.

Note: here we talk about SQLAlchemy collections used to manipulate
relationships, not db.Collection.

Closes https://github.com/chartes/lettres-app/issues/101