feat: put documents in the triage collection when and only when it has no other collection
It requires changing the loading strategy of document's collection to
something other than lazy in order to have up to date data in the
event's listener without requiring a session.flush. Indeed, maybe
because of a bug, flushing the session inside of an event listener seems
to not update the instances state and when we hit the `commit`
statement, later in `delete_related_resources`, `DELETE` sql statements
are executed a second time.
This leads to an error because sqlalchmey except the DELETE statement to
return one row receives none on commit (the data has already been
deleted from the DB).

We don't expect documents to be in a lot of collections so it shouldn't
impact performances.

Closes #108.