fix: "GET /languages" request not being handled correctly by flask
It seems that the werkzeug version we used had could have a bug when
building wsgi requests to be handled by flask when in multi-thread mode :

After removing a witness, when the document was reloaded
the `GET` request to the `/languages` route was transformed by werkzeug
which replaced the GET method by `{somejsonfrompreviousrequest}GET`.

Using other wsgi servers, adding threaded=False to flask_app.oy or
upgrading werkzeug seem to fix the problem.

Closes https://github.com/chartes/lettres-vue/issues/115